Move the SecondaryHierarchy to its own file in IBTK.
